+++ app/components/settings/menu/SettingsMenu.js
@@ -12,6 +12,11 @@ const messages = defineMessages({
     defaultMessage: '!!!General',
     description: 'Label for the "General" link in the settings menu.'
   },
+  termsOfUse: {
+    id: 'settings.menu.termsOfUse.link.label',
+    defaultMessage: '!!!Terms of use',
+    description: 'Label for the "Terms of use" link in the settings menu.'
+  },
 });
 
 @observer
@@ -37,6 +42,12 @@ export default class SettingsMenu extends Component {
             onClick={() => onItemClick(ROUTES.SETTINGS.GENERAL)}
             active={isActiveItem(ROUTES.SETTINGS.GENERAL)}
             className="general"
+          />
+          <SettingsMenuItem
+            label={intl.formatMessage(messages.termsOfUse)}
+            onClick={() => onItemClick(ROUTES.SETTINGS.TERMS_OF_USE)}
+            active={isActiveItem(ROUTES.SETTINGS.TERMS_OF_USE)}
+            className="termsOfUse"
           />
         </div>
       </div>